add one model one region model pipeline
